requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
neo4j>=5.20,<6.0
pydantic>=2.8,<3.0
python-dotenv>=1.0,<2.0
